Fixing what I hope are the last batch of problems caused by the addition of the suidmanager module.  Also fixing a couple of other small issues that somehow cropped up.  All tests should now pass again.
git-svn-id: https://reductivelabs.com/svn/puppet/trunk@1699 980ebf18-57e1-0310-9a29-db15c13687c0

+++ test/lib/puppettest.rb
@@ -151,6 +151,10 @@ module PuppetTest
         @@cleaners.each { |cleaner| cleaner.call() }
 
         @@tmpfiles.each { |file|
+            unless file =~ /tmp/
+                puts "Not deleting tmpfile %s" % file
+                next
+            end
             if FileTest.exists?(file)
                 system("chmod -R 755 %s" % file)
                 system("rm -rf %s" % file)
